Stop using typedef for enums and structs
A friend of mine pointed out that typedef'ing my enums and structs made it
impossible to know what kind of type he was dealing with when reading the code
source (without looking into other files, that is). So, this commit undoes the
typedef usage. Fun fact, this language is named after him! He's been a really
good friend to me, as well as a mentor in how to write good code.