Move struct implementations back to tokens.h file
I misunderstood an error that I was encountering with the compiler, and thought
that I had to fix the issue by moving the implementations of the token structs
into the tokens.c file. It turns out that is not the case, and I had only made
them opaque. I believe the issue I was running into was actually conflicting
enum identifiers. As such, the enums in question will soon be updated to not
be as prone to this type of conflict.