Fix segfault in tokens.c
I made a silly mistake in my tokens.c file, having NO_SYMB defined in the
symb_descs array and UNRESERVED defined in the rsvd_words array. Due to the
nature of the two, I set their symbol/identifier to NULL. As a result, when
time came for strcmp() in read_symb/read_word, there would be a segmentation
fault. I thought strcmp() knew how to handle NULL pointers, but I turned out
to be wrong. Yay for undefined behavior!